+++ extension/src/features/autofill/engine/index.ts
@@ -1,0 +1,6 @@
+export * from './DOMScanner';
+export * from './FieldMapper';
+export * from './ValueResolver';
+export * from './FormFiller';
+export * from './Logger';
+export * from './ResultReporter';